Moved ConnectionStringBuilder out of registry.py.
