+++ python/lsst/daf/butler/registries/sqlRegistry.py
@@ -41,7 +41,7 @@ from ..core.run import Run
 from ..core.storageClass import StorageClassFactory
 from ..core.config import Config
 from ..core.dimensions import DataId, Dimension
-from lsst.daf.butler.core.registry import ConnectionStringBuilder
+from lsst.daf.butler.core.connectionStringBuilder import ConnectionStringBuilder
 from .sqlRegistryDatabaseDict import SqlRegistryDatabaseDict
